Feature/publicspotify

+++ src/app/services/spotify/service.js
@@ -13,6 +13,8 @@ angular.module('spotmop.services.spotify', [])
 	// setup response object
     var service = {
 		
+		authenticationMethod: 'server',
+		
 		start: function(){
 	
 			// inject our authorization frame, on the placeholder action
@@ -52,7 +54,15 @@ angular.module('spotmop.services.spotify', [])
 				$localStorage.spotify.AccessToken = data.access_token;
 				$localStorage.spotify.RefreshToken = data.refresh_token;
 				$rootScope.spotifyOnline = true;
-				$rootScope.$broadcast('spotmop:spotify:online');
+				this.authenticationMethod = 'client';
+				
+				// get my details and store 'em
+				service.getMe()
+					.then( function(response){
+						SettingsService.setSetting('spotifyuser', response);
+					});
+				
+				$rootScope.$broadcast('spotmop:spotify:authenticationChanged', this.authenticationMethod);
 			}, false);
 			
 			
@@ -60,15 +70,22 @@ angular.module('spotmop.services.spotify', [])
 			 * The real starter
 			 **/
 			if( this.isAuthorized() ){
-				$rootScope.$broadcast('spotmop:spotify:online');
+				$rootScope.spotifyAuthorized = true;
+				this.authenticationMethod = 'client';
 			}else{
-				this.authorize();
+				SettingsService.setSetting('spotifyuser', false);
+				$rootScope.spotifyAuthorized = false;
+				this.authenticationMethod = 'server';
 			}
+			
+			$rootScope.$broadcast('spotmop:spotify:online');
 		},
 		
 		logout: function(){
 			$localStorage.spotify = {};
-			$rootScope.spotifyOnline = false;
+			this.authenticationMethod = 'server';
+			this.refreshToken();
+			$rootScope.$broadcast('spotmop:spotify:authenticationChanged', this.authenticationMethod);
 		},
 		
 		/**
@@ -87,15 +104,35 @@ angular.module('spotmop.services.spotify', [])
 		},
 		
 		/**
-		 * Refresh our existing credentials, by parsing our Authorization refresh_token
+		 * Refresh our access_token. There are two ways we can achieve this:
+		 *		1. Client has authorized Spotmop by "Connecting to Spotify" for maximum features
+		 *		2. Use config of backend to generate access_tokens for basic authentication
 		 **/
         refreshToken: function(){
 			
             var deferred = $q.defer();
-
+			var url = '';
+			
+			// sweet, client has authorized interface!
+			if( this.authenticationMethod == 'client' ){
+			
+				url = 'http://jamesbarnsley.co.nz/spotmop.php?action=refresh&refresh_token='+$localStorage.spotify.RefreshToken;
+			
+			// client hasn't authorized spotmop with spotify, so let's just use the backend
+			}else if( this.authenticationMethod == 'server' ){
+			
+				var mopidyhost = SettingsService.getSetting("mopidyhost", window.location.hostname);
+				var mopidyport = SettingsService.getSetting("mopidyport", "6680");
+				url = 'http://'+mopidyhost+':'+mopidyport+'/spotmop/auth';
+			
+			// no authentication method, so cannot refresh token!
+			}else{
+				return false;
+			}
+			
             $http({
 					method: 'GET',
-					url: 'http://jamesbarnsley.co.nz/spotmop.php?action=refresh&refresh_token='+$localStorage.spotify.RefreshToken,
+					url: url,
 					dataType: "json",
 					async: false,
 					timeout: 10000
@@ -188,6 +225,11 @@ angular.module('spotmop.services.spotify', [])
         getMe: function(){
 			
             var deferred = $q.defer();
+			
+			if( !this.isAuthorized() ){
+                deferred.reject();
+				return deferred.promise;
+			}
 
             $http({
 					method: 'GET',
@@ -211,16 +253,12 @@ angular.module('spotmop.services.spotify', [])
         
         getUser: function( useruri ){
 		
-			var userid = this.getFromUri( 'userid', useruri );
-			
+			var userid = this.getFromUri( 'userid', useruri );			
             var deferred = $q.defer();
 
             $http({
 					method: 'GET',
-					url: urlBase+'users/'+userid,
-					headers: {
-						Authorization: 'Bearer '+ $localStorage.spotify.AccessToken
-					}
+					url: urlBase+'users/'+userid
 				})
                 .success(function( response ){
 					
@@ -237,9 +275,13 @@ angular.module('spotmop.services.spotify', [])
 		
 		isFollowing: function( type, uri ){
 			
-			var id = this.getFromUri( type+'id', uri );
-			
+			var id = this.getFromUri( type+'id', uri );			
             var deferred = $q.defer();
+			
+			if( !this.isAuthorized() ){
+                deferred.reject();
+				return deferred.promise;
+			}
 
             $http({
 					method: 'GET',
@@ -268,16 +310,12 @@ angular.module('spotmop.services.spotify', [])
 	
 		getTrack: function( trackuri ){
 			
-			var trackid = this.getFromUri('trackid', trackuri);
-			
+			var trackid = this.getFromUri('trackid', trackuri);			
             var deferred = $q.defer();
 
             $http({
 					method: 'GET',
-					url: urlBase+'tracks/'+trackid,
-					headers: {
-						Authorization: 'Bearer '+ $localStorage.spotify.AccessToken
-					}
+					url: urlBase+'tracks/'+trackid
 				})
                 .success(function( response ){
 					
@@ -301,6 +339,11 @@ angular.module('spotmop.services.spotify', [])
 		getMyTracks: function( userid ){
 			
             var deferred = $q.defer();
+			
+			if( !this.isAuthorized() ){
+                deferred.reject();
+				return deferred.promise;
+			}
 
             $http({
 					method: 'GET',
@@ -325,6 +368,11 @@ angular.module('spotmop.services.spotify', [])
 		addTracksToLibrary: function( trackids ){
 			
             var deferred = $q.defer();
+			
+			if( !this.isAuthorized() ){
+                deferred.reject();
+				return deferred.promise;
+			}
 
             $http({
 					method: 'PUT',
@@ -352,6 +400,11 @@ angular.module('spotmop.services.spotify', [])
 		deleteTracksFromLibrary: function( trackids ){
 			
             var deferred = $q.defer();
+			
+			if( !this.isAuthorized() ){
+                deferred.reject();
+				return deferred.promise;
+			}
 
             $http({
 					method: 'DELETE',
@@ -379,6 +432,11 @@ angular.module('spotmop.services.spotify', [])
 		getMyArtists: function( userid ){
 			
             var deferred = $q.defer();
+			
+			if( !this.isAuthorized() ){
+                deferred.reject();
+				return deferred.promise;
+			}
 
             $http({
 					method: 'GET',
@@ -402,9 +460,13 @@ angular.module('spotmop.services.spotify', [])
 		
 		isFollowingArtist: function( artisturi, userid ){
 			
-			var artistid = this.getFromUri( 'artistid', artisturi );			
-			
+			var artistid = this.getFromUri( 'artistid', artisturi );
             var deferred = $q.defer();
+			
+			if( !this.isAuthorized() ){
+                deferred.reject();
+				return deferred.promise;
+			}
 
             $http({
 					cache: false,
@@ -427,9 +489,13 @@ angular.module('spotmop.services.spotify', [])
 		
 		followArtist: function( artisturi ){
 			
-			var artistid = this.getFromUri( 'artistid', artisturi );			
-			
+			var artistid = this.getFromUri( 'artistid', artisturi );	
             var deferred = $q.defer();
+			
+			if( !this.isAuthorized() ){
+                deferred.reject();
+				return deferred.promise;
+			}
 
             $http({
 					method: 'PUT',
@@ -452,9 +518,13 @@ angular.module('spotmop.services.spotify', [])
 		
 		unfollowArtist: function( artisturi ){
 			
-			var artistid = this.getFromUri( 'artistid', artisturi );			
-			
+			var artistid = this.getFromUri( 'artistid', artisturi );
             var deferred = $q.defer();
+			
+			if( !this.isAuthorized() ){
+                deferred.reject();
+				return deferred.promise;
+			}
 
             $http({
 					method: 'DELETE',
@@ -485,7 +555,6 @@ angular.module('spotmop.services.spotify', [])
 			if( typeof( limit ) === 'undefined' )
 				limit = 40;
 				
-			
             var deferred = $q.defer();
 
             $http({
@@ -512,8 +581,6 @@ angular.module('spotmop.services.spotify', [])
 			// get the user and playlist ids from the uri
 			var userid = this.getFromUri( 'userid', playlisturi );
 			var playlistid = this.getFromUri( 'playlistid', playlisturi );
-				
-			
             var deferred = $q.defer();
 
             $http({
@@ -539,9 +606,12 @@ angular.module('spotmop.services.spotify', [])
 			
 			var userid = this.getFromUri( 'userid', playlisturi );
 			var playlistid = this.getFromUri( 'playlistid', playlisturi );
-			
-			
             var deferred = $q.defer();
+			
+			if( !this.isAuthorized() ){
+                deferred.reject();
+				return deferred.promise;
+			}
 
             $http({
 					cache: true,
@@ -566,9 +636,12 @@ angular.module('spotmop.services.spotify', [])
 			
 			var userid = this.getFromUri( 'userid', playlisturi );
 			var playlistid = this.getFromUri( 'playlistid', playlisturi );
-			
-			
             var deferred = $q.defer();
+			
+			if( !this.isAuthorized() ){
+                deferred.reject();
+				return deferred.promise;
+			}
 
             $http({
 					method: 'PUT',
@@ -592,9 +665,12 @@ angular.module('spotmop.services.spotify', [])
 			
 			var userid = this.getFromUri( 'userid', playlisturi );
 			var playlistid = this.getFromUri( 'playlistid', playlisturi );
-			
-			
             var deferred = $q.defer();
+			
+			if( !this.isAuthorized() ){
+                deferred.reject();
+				return deferred.promise;
+			}
 
             $http({
 					method: 'DELETE',
@@ -620,9 +696,7 @@ angular.module('spotmop.services.spotify', [])
 				limit = 40;
 			
 			var timestamp = $filter('date')(new Date(),'yyyy-MM-ddTHH:mm:ss');
-			var country = SettingsService.getSetting('countrycode','NZ');
-			
-			
+			var country = SettingsService.getSetting('countrycode','NZ');						
             var deferred = $q.defer();
 
             $http({
@@ -650,8 +724,12 @@ angular.module('spotmop.services.spotify', [])
 			var userid = this.getFromUri( 'userid', playlisturi );
 			var playlistid = this.getFromUri( 'playlistid', playlisturi );
 			
-			
             var deferred = $q.defer();
+			
+			if( !this.isAuthorized() ){
+                deferred.reject();
+				return deferred.promise;
+			}
 
             $http({
 					method: 'POST',
@@ -683,6 +761,10 @@ angular.module('spotmop.services.spotify', [])
             if( userid != SettingsService.getSetting('spotifyuserid',null) )
                 return false;
 			
+			if( !this.isAuthorized() ){
+                deferred.reject();
+				return deferred.promise;
+			}			
 			
             var deferred = $q.defer();
 
@@ -745,6 +827,11 @@ angular.module('spotmop.services.spotify', [])
 		createPlaylist: function( userid, data ){
 			
             var deferred = $q.defer();
+			
+			if( !this.isAuthorized() ){
+                deferred.reject();
+				return deferred.promise;
+			}
 
             $http({
 					method: 'POST',
@@ -777,6 +864,11 @@ angular.module('spotmop.services.spotify', [])
 			var playlistid = this.getFromUri( 'playlistid', playlisturi );
 			
             var deferred = $q.defer();
+			
+			if( !this.isAuthorized() ){
+                deferred.reject();
+				return deferred.promise;
+			}
 
             $http({
 					method: 'PUT',
@@ -916,10 +1008,7 @@ angular.module('spotmop.services.spotify', [])
             $http({
 					cache: true,
 					method: 'GET',
-					url: urlBase+'artists/'+artistid,
-					headers: {
-						Authorization: 'Bearer '+ $localStorage.spotify.AccessToken
-					}
+					url: urlBase+'artists/'+artistid
 				})
                 .success(function( response ){					
                     deferred.resolve( response );
@@ -945,10 +1034,7 @@ angular.module('spotmop.services.spotify', [])
 
             $http({
 					method: 'GET',
-					url: urlBase+'artists/?ids='+artistids,
-					headers: {
-						Authorization: 'Bearer '+ $localStorage.spotify.AccessToken
-					}
+					url: urlBase+'artists/?ids='+artistids
 				})
                 .success(function( response ){					
                     deferred.resolve( response );
@@ -969,10 +1055,7 @@ angular.module('spotmop.services.spotify', [])
             $http({
 					cache: true,
 					method: 'GET',
-					url: urlBase+'artists/'+artistid+'/albums?album_type=album,single&market='+country,
-					headers: {
-						Authorization: 'Bearer '+ $localStorage.spotify.AccessToken
-					}
+					url: urlBase+'artists/'+artistid+'/albums?album_type=album,single&market='+country
 				})
                 .success(function( response ){					
                     deferred.resolve( response );
@@ -992,10 +1075,7 @@ angular.module('spotmop.services.spotify', [])
 
             $http({
 					method: 'GET',
-					url: urlBase+'albums/'+albumid,
-					headers: {
-						Authorization: 'Bearer '+ $localStorage.spotify.AccessToken
-					}
+					url: urlBase+'albums/'+albumid
 				})
                 .success(function( response ){					
                     deferred.resolve( response );
@@ -1016,10 +1096,7 @@ angular.module('spotmop.services.spotify', [])
             $http({
 					cache: true,
 					method: 'GET',
-					url: urlBase+'artists/'+artistid+'/top-tracks?country='+country,
-					headers: {
-						Authorization: 'Bearer '+ $localStorage.spotify.AccessToken
-					}
+					url: urlBase+'artists/'+artistid+'/top-tracks?country='+country
 				})
                 .success(function( response ){
                     deferred.resolve( response );
@@ -1040,10 +1117,7 @@ angular.module('spotmop.services.spotify', [])
             $http({
 					cache: true,
 					method: 'GET',
-					url: urlBase+'artists/'+artistid+'/related-artists',
-					headers: {
-						Authorization: 'Bearer '+ $localStorage.spotify.AccessToken
-					}
+					url: urlBase+'artists/'+artistid+'/related-artists'
 				})
                 .success(function( response ){
                     deferred.resolve( response );
@@ -1139,40 +1213,23 @@ angular.module('spotmop.services.spotify', [])
 			if( response.status == 401 && response.config.url.search('https://api.spotify.com/') >= 0 && retryCount < 3 ){
 					
 				retryCount++;
-				var deferred = $q.defer();				
+				var deferred = $q.defer();
 				
-				// if we're already authorized, we just need to force a token refresh
-				if( $injector.get('SpotifyService').isAuthorized() ){
+				// refresh the token
+				$injector.get('SpotifyService').refreshToken()
+					.then( function(refreshResponse){
+						
+						// make sure our refresh request didn't error, otherwise we'll create an infinite loop
+						if( typeof(refreshResponse.error) !== 'undefined' )
+							return response;
+							
+						retryCount--;
+						
+						// now retry the original request
+						retryHttpRequest( response.config, deferred, refreshResponse.access_token );
+					});
 				
-					// refresh the token
-					$injector.get('SpotifyService').refreshToken()
-						.then( function(refreshResponse){
-							
-							// make sure our refresh request didn't error, otherwise we'll create an infinite loop
-							if( typeof(refreshResponse.error) !== 'undefined' )
-								return response;
-								
-							retryCount--;
-							
-							// now retry the original request
-							retryHttpRequest( response.config, deferred, refreshResponse.access_token );
-						});
-					
-					return deferred.promise;
-		
-				// not yet authorized, so authorize!
-				// this requires user interaction, so let's just allow the original request to fail
-				}else{
-					
-					// remove our current authorization, just to clear the decks
-					$localStorage.spotify = {};
-					$rootScope.spotifyOnline = false;
-					
-					// and re-authorize
-					$injector.get('SpotifyService').authorize();	
-					retryCount--;
-					return response;
-				}
+				return deferred.promise;
 			}
 			
 			return response;
